WC Blocks: Refactor sponsor blocks (see desc)
Ideally this should multiple separate commits, but its not trivial to seperate out so making this in a single commit. This has following changes:

1. Refactor Sponsor blocks so thats its mostly consistent with rest of the block structure. Towards https://github.com/WordPress/wordcamp.org/issues/57

2. Removed `componentWillReceiveProps` because its being deprecated. Instead used `getDerivedStateFromProps` to memoize sponsor post order. Towards https://github.com/WordPress/wordcamp.org/issues/63

3. Refactor to use blocks store instead of manual `apiFetch` calls.
